April 20, 2017
Added onClick attribute to xml, made Edit Text variables member
variables, made a createUser method in MainActivity class, we made
profileUser of type User and we assigned it to the constructor. We gave
it the parameters of emailEditText and passwordEditText which converted
to Strings. Then created a toast message to show the user's input for
email and password. Also changed the app name.