April 20, 2017
-added android on click attribute to xml
- we made our editText variables to member variables
-we made profileUser of type User and we assigned it to the constructor,
we gave it the perameters of email edit text and password edit text. We
than called methods on those perameters to convert the parameters to
strings. We followed with a toast message, it displays a short message.
we got it to show by putting in the "show" method after the parameters.