April 22
We created an object called user1 which would give us the username and
the password, which would then be stored into the variable we created
known as userStorage. We also created a toast message that would display
the username and password. We created a toast message for each scenario
; if the user had already been selected or if a new user  was entered.
We then created a listener object so that we would no longer need the
onClick method in the xml. Instead we created the the Listener object to
use the onClick method to run the program.